+++ tests/__init__.py
@@ -1,2 +1,7 @@
 from .test_create import *
 from .test_parse import *
+from .test_encoding_error import *
+from .test_encode import *
+from .test_decode import *
+from .test_hash_field import *
+from .test_hash_raw import *